chore(test): address PR review comments for ValidateObjects tests
- Move ValidateObjects tests to per-type test files (trainingruntime_test.go and clustertrainingruntime_test.go) instead of standalone validateobjects_test.go
- Use cmp.Diff pattern for assertions instead of manual field error counting
- Remove duplicate test case for runtime in different namespace
- Simplify warnings handling by using []string type conversion

This addresses review comments from @robert-bell to improve test organization and consistency with existing test patterns in the codebase.
